Fixed styling issues. Renamed the default tooltip to DefaultTooltip. Removed consol logs
