+++ src/ButtonPanel.css
@@ -58,7 +58,6 @@
   border: 1px solid #474949;
   background-image: linear-gradient(#4F5151, #474949 6%, #3F4141);
   background-color: white;
-  line-height: 1;
   text-align: center;
   vertical-align: middle;
   cursor: pointer;
