+++ src/Dropdown.css
@@ -9,7 +9,7 @@
   border: 1px solid #ddd;
   display: inline-flex;
   box-sizing: border-box;
-  width: 410px;
+  width: 310px;
 }
 
 .escher-container .dropdown {
@@ -25,7 +25,6 @@
   -ms-touch-action: none;
   touch-action: none;
   cursor: pointer;
-  font-family: "Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;
   -webkit-touch-callout: none;
   -webkit-user-select: none;
   -khtml-user-select: none;
@@ -41,15 +40,15 @@
 }
 
 .escher-container .helpButton {
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif!important;
   color: #d9230f!important;
-  padding: 2px 5px;
+  padding: 0px 5px;
   background-color: #f5f5f5!important;
   border: 0!important;
   text-decoration: unset;
   height: fit-content;
   position: absolute;
   right: 0;
+  font-size: 12px;
 }
 
 .escher-container .dropdownButton:focus, .escher-container .helpButton:focus {
@@ -63,6 +62,9 @@
 }
 
 @media (min-width: 550px) {
+  .escher-container .menuBar {
+    width: 410px;
+  }
   .escher-container .dropdown {
     font-size: 18px;
   }
@@ -74,10 +76,6 @@
   .escher-container .helpButton {
     font-size: 16px;
   }
-}
-
-.escher-container .helpButton {
-  font-size: 16px;
 }
 
 .escher-container .menu {
@@ -118,7 +116,6 @@
   display: block;
   font-weight: normal!important;
   color: #444444;
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif !important;
 }
 
 .escher-container .menuBar {
